PATCH tests do not extend PUT tests
