+++ rest/admin-v2/tests/src/test/java/org/keycloak/tests/admin/client/v2/validation/AbstractClientValidationTest.java
@@ -24,8 +24,12 @@ import org.keycloak.common.Profile;
 import org.keycloak.representations.admin.v2.OIDCClientRepresentation;
 import org.keycloak.representations.admin.v2.SAMLClientRepresentation;
 import org.keycloak.services.error.ViolationExceptionResponse;
+import org.keycloak.testframework.annotations.InjectClient;
 import org.keycloak.testframework.annotations.InjectHttpClient;
 import org.keycloak.testframework.annotations.InjectRealm;
+import org.keycloak.testframework.realm.ClientConfig;
+import org.keycloak.testframework.realm.ClientConfigBuilder;
+import org.keycloak.testframework.realm.ManagedClient;
 import org.keycloak.testframework.realm.ManagedRealm;
 import org.keycloak.testframework.server.KeycloakServerConfig;
 import org.keycloak.testframework.server.KeycloakServerConfigBuilder;
@@ -64,6 +68,12 @@ public abstract class AbstractClientValidationTest extends AbstractClientApiV2Te
 
     @InjectRealm
     ManagedRealm realm;
+
+    @InjectClient(config = PutClientValidationTest.TestOidcClient.class, ref = "test-oidc")
+    ManagedClient testOidcClient;
+
+    @InjectClient(config = PutClientValidationTest.TestSamlClient.class, ref = "test-saml")
+    ManagedClient testSamlClient;
 
     @InjectHttpClient
     CloseableHttpClient client;
@@ -501,6 +511,28 @@ public abstract class AbstractClientValidationTest extends AbstractClientApiV2Te
         }
     }
 
+    @ParameterizedTest
+    @ValueSource(strings = {OIDCClientRepresentation.PROTOCOL, SAMLClientRepresentation.PROTOCOL})
+    public void clientIdMismatchBetweenPathAndPayloadFails(String protocol) throws Exception {
+        boolean isOidc = protocol.equals(OIDCClientRepresentation.PROTOCOL);
+        var request = getRequest(isOidc);
+        setAuthHeader(request);
+
+        request.setEntity(new StringEntity("""
+                {
+                    "protocol": "%s",
+                    "clientId": "different-client-id",
+                    "enabled": true
+                }
+                """.formatted(protocol)));
+
+        try (var response = client.execute(request)) {
+            assertThat(response.getStatusLine().getStatusCode(), is(400));
+            var body = mapper.createParser(response.getEntity().getContent()).readValueAs(ViolationExceptionResponse.class);
+            assertThat(body.error(), is("Field 'clientId' in payload does not match the provided 'clientId'"));
+        }
+    }
+
     protected boolean isPutRequest() {
         return getHttpMethod().equals(HttpPut.METHOD_NAME);
     }
@@ -509,6 +541,24 @@ public abstract class AbstractClientValidationTest extends AbstractClientApiV2Te
         @Override
         public KeycloakServerConfigBuilder configure(KeycloakServerConfigBuilder config) {
             return config.features(Profile.Feature.CLIENT_ADMIN_API_V2);
+        }
+    }
+
+    public static class TestOidcClient implements ClientConfig {
+        @Override
+        public ClientConfigBuilder configure(ClientConfigBuilder client) {
+            return client.clientId("test-client-oidc")
+                    .enabled(true)
+                    .protocol("openid-connect");
+        }
+    }
+
+    public static class TestSamlClient implements ClientConfig {
+        @Override
+        public ClientConfigBuilder configure(ClientConfigBuilder client) {
+            return client.clientId("test-client-saml")
+                    .enabled(true)
+                    .protocol("saml");
         }
     }
 }
